+++ pycvvdp/run_cvvdp.py
@@ -105,7 +105,7 @@ def parse_args(arg_list=None):
     parser.add_argument("-q", "--quiet", action='store_true', default=False, help="Do not print any information but the final JOD value. Warning message will be still printed.")
     parser.add_argument("-v", "--verbose", action='store_true', default=False, help="Print out extra information.")
     parser.add_argument("--ffmpeg-cc", action='store_true', default=False, help="Use ffmpeg for upsampling and color conversion. Use custom pytorch code by default (faster and less memory).")    
-    parser.add_argument("--temp-resample", action='store_true', default=False, help="Resample test and reference video to a common frame rate. Allows to compare videos of different frame rates.")
+    parser.add_argument("--temp-resample", type=float, nargs="?", default=-1, const=0, help="Resample test and reference video to a common frame rate. Allows to compare videos of different frame rates. An optional argument - the maximum frame rate used when resampling.")
     parser.add_argument("-i", "--interactive", action='store_true', default=False, help="Run in an interactive mode, in which command line arguments are provided to the standard input, line by line. Saves on start-up time when running a large number of comparisons.")
     parser.add_argument("--dump-channels", nargs='+', choices=['temporal', 'lpyr', 'difference'], default=None, help="Output video/images with intermediate processing stages (for debugging and visualization).")
     if arg_list is not None:
@@ -291,7 +291,9 @@ def run_on_args(args):
             preload = False if args.temp_padding == 'replicate' else True
             with torch.no_grad():
 
-                if args.temp_resample:
+                if args.temp_resample>=0:
+                    if args.temp_resample>0:
+                        pycvvdp.video_source_temp_resample_file.max_fps = args.temp_resample
                     vs = pycvvdp.video_source_temp_resample_file( test_file, ref_file, 
                                                 display_photometry=display_photometry, 
                                                 config_paths=args.config_paths,
@@ -318,13 +320,14 @@ def run_on_args(args):
                 mm.set_base_fname(base_fname)
 
                 Q_pred, stats = mm.predict_video_source(vs)
+                Q_pred_scalar = Q_pred.item()
                 if args.quiet:
-                    print( "{Q:0.4f}".format(Q=Q_pred) )
+                    print( "{Q:0.4f}".format(Q=Q_pred_scalar) )                    
                 else:
                     units_str = f" [{mm.quality_unit()}]"
-                    print( "{met_name}={Q:0.4f}{units}".format(met_name=mm.short_name(), Q=Q_pred, units=units_str) )
+                    print( "{met_name}={Q:0.4f}{units}".format(met_name=mm.short_name(), Q=Q_pred_scalar, units=units_str) )
                 if not res_fh is None:
-                    res_fh.write( f", {Q_pred}" )
+                    res_fh.write( f", {Q_pred_scalar}" )
 
 
                 if args.features and not stats is None:
@@ -345,14 +348,11 @@ def run_on_args(args):
                         logging.info("Writing heat map '" + dest_name + "' ...")
                         np2img(torch.squeeze(stats["heatmap"].permute((2,3,4,1,0)), dim=4).cpu().numpy(), dest_name)
 
-                if args.distogram != -1 and not stats is None:
+                if args.distogram != -1:
                     dest_name = os.path.join(out_dir, base + "_distogram.png")                    
                     logging.info("Writing distogram '" + dest_name + "' ...")
                     jod_max = args.distogram
-                    try: 
-                        mm.export_distogram( stats, dest_name, jod_max=jod_max )
-                    except NotImplementedError as e:
-                        logging.warning( f'Metric {mm.short_name()} cannot generate distograms' )
+                    mm.export_distogram( stats, dest_name, jod_max=jod_max )
                     
                 del stats
 
@@ -368,19 +368,21 @@ def run_on_args(args):
 def main():
     args = parse_args()
 
-    if args.interactive:
-        #print( "Running in an interactive mode" )
-        while True:
-            line = sys.stdin.readline()
-            if not line:
-                break
+    try:
+        if args.interactive:
+            #print( "Running in an interactive mode" )
+            while True:
+                line = sys.stdin.readline()
+                if not line:
+                    break
 
-            #print( shlex.split(line) )
-            args = parse_args(shlex.split(line))
+                #print( shlex.split(line) )
+                args = parse_args(shlex.split(line))
+                run_on_args(args)
+        else:
             run_on_args(args)
-    else:
-        run_on_args(args)
-
+    except pycvvdp.vq_exception as ex:
+        logging.error( str(ex) )
 
 if __name__ == '__main__':
     main()